Renamed folders and route
