+++ backend/src/modules/index.ts
@@ -1,1 +1,2 @@
-export * from "./auth";
+export { router as authRouter } from "./auth";
+export { router as rabbitRouter } from "./rabbit";
