+++ backend/src/modules/rabbit/index.ts
